Move ViewModel namespace to ViewModels.
