+++ MinimalMVVM/ViewModels/Presenter.cs
@@ -3,7 +3,7 @@ using System.Collections.ObjectModel;
 using System.Windows.Input;
 using MinimalMVVM.Models;
 
-namespace MinimalMVVM.ViewModel
+namespace MinimalMVVM.ViewModels
 {
     public class Presenter : ObservableObject
     {
