+++ MinimalMVVM/Views/ConvertWindow.xaml
@@ -1,15 +1,15 @@
 <Window x:Class="MinimalMVVM.Views.ConvertWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:ViewModel="clr-namespace:MinimalMVVM.ViewModel"
         xmlns:Views="clr-namespace:MinimalMVVM.Views"
+        xmlns:ViewModels="clr-namespace:MinimalMVVM.ViewModels"
         Title="Converter"
         MinWidth="300"
         ResizeMode="NoResize"
         SizeToContent="WidthAndHeight">
     
     <Window.DataContext>
-        <ViewModel:Presenter/>
+        <ViewModels:Presenter/>
     </Window.DataContext>
     
     <Views:ConverterControl/>
