+++ MinimalMVVM/Views/ConverterControl.xaml
@@ -3,10 +3,11 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:ViewModel="clr-namespace:MinimalMVVM.ViewModel" mc:Ignorable="d" 
+             xmlns:ViewModels="clr-namespace:MinimalMVVM.ViewModels"
+             mc:Ignorable="d" 
              d:DesignHeight="336"
              d:DesignWidth="300"
-             d:DataContext="{d:DesignInstance ViewModel:Presenter}">
+             d:DataContext="{d:DesignInstance ViewModels:Presenter}">
 
     <UserControl.InputBindings>
         <KeyBinding Key="Enter" Command="{Binding ConvertTextCommand}"/>
